Fix keyboard hiding when closing fragments.
