+++ Android/ghvzApp/.idea/navEditor.xml
@@ -78,7 +78,12 @@
                       <LayoutPositions />
                     </value>
                   </entry>
-                  <entry key="action_global_nav_quiz_question_order_fragment">
+                  <entry key="action_global_nav_quiz_question_multi_answer_fragment">
+                    <value>
+                      <LayoutPositions />
+                    </value>
+                  </entry>
+                  <entry key="action_global_nav_quiz_question_true_false_fragment">
                     <value>
                       <LayoutPositions />
                     </value>
@@ -298,7 +303,7 @@
                       </LayoutPositions>
                     </value>
                   </entry>
-                  <entry key="nav_quiz_question_order_fragment">
+                  <entry key="nav_quiz_question_multi_answer_fragment">
                     <value>
                       <LayoutPositions>
                         <option name="myPosition">
@@ -310,12 +315,12 @@
                       </LayoutPositions>
                     </value>
                   </entry>
-                  <entry key="nav_quiz_question_t_fragment">
+                  <entry key="nav_quiz_question_true_false_fragment">
                     <value>
                       <LayoutPositions>
                         <option name="myPosition">
                           <Point>
-                            <option name="x" value="12" />
+                            <option name="x" value="256" />
                             <option name="y" value="12" />
                           </Point>
                         </option>
