+++ Android/ghvzApp/app/src/main/java/com/app/playhvz/screens/missions/missionsettings/MissionSettingsFragment.kt
@@ -125,7 +125,7 @@ class MissionSettingsFragment : Fragment() {
     }
 
     override fun onDestroyView() {
-        SystemUtils.hideKeyboard(requireContext())
+        SystemUtils.hideKeyboard(requireView())
         super.onDestroyView()
     }
 
@@ -153,6 +153,11 @@ class MissionSettingsFragment : Fragment() {
             submitMission()
         }
         return super.onOptionsItemSelected(item)
+    }
+
+    override fun onPause() {
+        super.onPause()
+        SystemUtils.hideKeyboard(requireView())
     }
 
     private fun setupObservers() {
@@ -304,7 +309,7 @@ class MissionSettingsFragment : Fragment() {
     }
 
     private fun disableActions() {
-        SystemUtils.hideKeyboard(requireContext())
+        SystemUtils.hideKeyboard(requireView())
         val menuItem = toolbarMenu.findItem(R.id.save_option)
         menuItem.icon.mutate().alpha = 130
         menuItem.isEnabled = false
