+++ Android/ghvzApp/app/src/main/java/com/app/playhvz/screens/quiz/editablequestions/BooleanAnswerAdapter.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.app.playhvz.screens.quiz.questions
+package com.app.playhvz.screens.quiz.editablequestions
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
