+++ Android/ghvzApp/app/src/main/java/com/app/playhvz/screens/quiz/editablequestions/TrueFalseQuestionFragment.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.app.playhvz.screens.quiz.questions
+package com.app.playhvz.screens.quiz.editablequestions
 
 import android.os.Bundle
 import android.view.*
@@ -120,7 +120,7 @@ class TrueFalseQuestionFragment : Fragment() {
 
     override fun onPause() {
         super.onPause()
-        SystemUtils.hideKeyboard(requireContext())
+        SystemUtils.hideKeyboard(requireView())
     }
 
     override fun onResume() {
@@ -177,7 +177,7 @@ class TrueFalseQuestionFragment : Fragment() {
     }
 
     private fun disableActions() {
-        SystemUtils.hideKeyboard(requireContext())
+        SystemUtils.hideKeyboard(requireView())
         val menuItem = toolbarMenu.findItem(R.id.save_option)
         menuItem.icon.mutate().alpha = 130
         menuItem.isEnabled = false
